refactor: ♻️ Expose read_properties as a public function
